Move all state population and general functions to a base class
