+++ src/game_api/CMakeLists.txt
@@ -47,6 +47,10 @@ add_library(spel2_api STATIC
         window_api.hpp
         script.cpp
         script.hpp
+        script/lua_vm.cpp
+        script/lua_vm.hpp
+        script/lua_backend.cpp
+        script/lua_backend.hpp
         script/script_impl.cpp
         script/script_impl.hpp
         script/script_util.cpp
