+++ src/game_api/script/usertypes/entities_floors_lua.cpp
@@ -8,7 +8,7 @@
 
 namespace NEntitiesFloors
 {
-void register_usertypes(sol::state& lua, ScriptImpl* script)
+void register_usertypes(sol::state& lua, LuaBackend* script)
 {
     lua["Entity"]["as_floor"] = &Entity::as<Floor>;
     lua["Entity"]["as_door"] = &Entity::as<Door>;
