+++ src/game_api/script/usertypes/entities_liquids_lua.hpp
@@ -2,9 +2,9 @@
 
 #include <sol/forward.hpp>
 
-class ScriptImpl;
+class LuaBackend;
 
 namespace NEntitiesLiquids
 {
-void register_usertypes(sol::state& lua, ScriptImpl* script);
+void register_usertypes(sol::state& lua, LuaBackend* script);
 };
