+++ src/game_api/script/usertypes/entity_casting_lua.cpp
@@ -6,7 +6,7 @@
 
 namespace NEntityCasting
 {
-void register_usertypes(sol::state& lua, ScriptImpl* script)
+void register_usertypes(sol::state& lua, LuaBackend* script)
 {
     lua.create_named_table("TYPE_MAP");
