+++ src/game_api/script/usertypes/level_lua.hpp
@@ -2,7 +2,7 @@
 
 #include <sol/forward.hpp>
 
-class ScriptImpl;
+class LuaBackend;
 
 using ROOM_TEMPLATE = uint16_t;     // NoAlias
 using PROCEDURAL_CHANCE = uint32_t; // NoAlias
@@ -19,5 +19,5 @@ struct PostRoomGenerationContext
 
 namespace NLevel
 {
-void register_usertypes(sol::state& lua, ScriptImpl* script);
+void register_usertypes(sol::state& lua, LuaBackend* script);
 };
