+++ src/game_api/script/usertypes/sound_lua.hpp
@@ -2,10 +2,10 @@
 
 #include <sol/forward.hpp>
 
-class ScriptImpl;
+class LuaBackend;
 class SoundManager;
 
 namespace NSound
 {
-void register_usertypes(sol::state& lua, ScriptImpl* script);
+void register_usertypes(sol::state& lua, LuaBackend* script);
 };
