+++ src/game_api/window_api.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cstdint>
 
 #include <Windows.h>
